backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "E-Business Backend API"
authors = ["Admin <admin@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
sqlalchemy = "^2.0.25"
asyncpg = "^0.29.0"
python-socketio = {extras = ["client"], version = "^5.16.0"}
aiohttp = "^3.9.0"
pydantic-settings = "^2.1.0"
alembic = "^1.13.1"
python-multipart = "^0.0.9"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
langgraph = "^0.2.0"
langchain-core = "^0.3.0"
email-validator = "^2.3.0"
requests = "^2.32.5"
python-dotenv = "^1.2.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.5"
pytest-cov = "^4.1.0"
httpx = "^0.26.0"
aiosqlite = "^0.19.0"
python-socketio = {extras = ["client"], version = "^5.16.0"}
requests = "^2.32.5"

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
